+++ src/main/java/org/bulatnig/smpp/pdu/AlertNotification.java
@@ -85,7 +85,7 @@ public class AlertNotification extends PDU {
         }
         SmppByteBuffer bb = new SmppByteBuffer(bytes);
         try {
-            short b = bb.removeByte();
+            int b = bb.removeByte();
             for (TON ton : TON.values()) {
                 if (ton.getValue() == b) {
                     sourceAddrTon = ton;
