+++ src/main/java/org/bulatnig/smpp/pdu/SubmitSM.java
@@ -73,11 +73,11 @@ public class SubmitSM extends PDU implements Responsable {
     /**
      * Protocol Identifier. Network specific field.
      */
-    private short protocolId;
+    private int protocolId;
     /**
      * Designates the priority level of the message.
      */
-    private short priorityFlag;
+    private int priorityFlag;
     /**
      * The short message is to be scheduled by the SMSC for delivery. Set to
      * NULL for immediate message delivery.
@@ -92,25 +92,25 @@ public class SubmitSM extends PDU implements Responsable {
      * Indicator to signify if an SMSC delivery receipt or an SME
      * acknowledgement is required.
      */
-    private short registeredDelivery;
+    private int registeredDelivery;
     /**
      * Flag indicating if submitted message should replace an existing message.
      */
-    private short replaceIfPresentFlag;
+    private int replaceIfPresentFlag;
     /**
      * Defines the encoding scheme of the short message user data.
      */
-    private short dataCoding;
+    private int dataCoding;
     /**
      * Indicates the short message to send from a list of predefined (‘canned’)
      * short messages stored on the SMSC. If not using an SMSC canned message,
      * set to NULL.
      */
-    private short smDefaultMsgId;
+    private int smDefaultMsgId;
     /**
      * Length in octets of the short_message user data.
      */
-    private short smLength;
+    private int smLength;
     /**
      * Up to 254 octets of short message user data. The exact physical limit for
      * short_message size may vary according to the underlying network.<br/>
@@ -431,7 +431,7 @@ public class SubmitSM extends PDU implements Responsable {
             if (serviceType.length() > MAX_SERVICETYPE_LENGTH) {
                 throw new PDUException("serviceType field is too long");
             }
-            short b = bb.removeByte();
+            int b = bb.removeByte();
             for (TON ton : TON.values()) {
                 if (ton.getValue() == b) {
                     sourceAddrTon = ton;
@@ -714,7 +714,7 @@ public class SubmitSM extends PDU implements Responsable {
     /**
      * @return Protocol Identifier
      */
-    public final short getProtocolId() {
+    public final int getProtocolId() {
         return protocolId;
     }
 
@@ -728,7 +728,7 @@ public class SubmitSM extends PDU implements Responsable {
     /**
      * @return the priority level of the message
      */
-    public final short getPriorityFlag() {
+    public final int getPriorityFlag() {
         return priorityFlag;
     }
 
@@ -771,7 +771,7 @@ public class SubmitSM extends PDU implements Responsable {
      * @return indicator to signify if an SMSC delivery receipt or an SME
      *         acknowledgement is required
      */
-    public final short getRegisteredDelivery() {
+    public final int getRegisteredDelivery() {
         return registeredDelivery;
     }
 
@@ -786,7 +786,7 @@ public class SubmitSM extends PDU implements Responsable {
     /**
      * @return if submitted message should replace an existing message indicator
      */
-    public final short getReplaceIfPresentFlag() {
+    public final int getReplaceIfPresentFlag() {
         return replaceIfPresentFlag;
     }
 
@@ -801,7 +801,7 @@ public class SubmitSM extends PDU implements Responsable {
     /**
      * @return encoding scheme of the short message user data
      */
-    public final short getDataCoding() {
+    public final int getDataCoding() {
         return dataCoding;
     }
 
@@ -816,7 +816,7 @@ public class SubmitSM extends PDU implements Responsable {
      * @return the short message to send from a list of predefined (‘canned’)
      *         short messages stored on the SMSC indicator
      */
-    public final short getSmDefaultMsgId() {
+    public final int getSmDefaultMsgId() {
         return smDefaultMsgId;
     }
 
@@ -831,7 +831,7 @@ public class SubmitSM extends PDU implements Responsable {
     /**
      * @return length in octets of the short_message user data
      */
-    public final short getSmLength() {
+    public final int getSmLength() {
         return smLength;
     }
