+++ src/main/java/org/bulatnig/smpp/pdu/UnsuccessSme.java
@@ -49,7 +49,7 @@ public class UnsuccessSme {
     public UnsuccessSme(final byte[] bytes) throws PDUException {
         SmppByteBuffer bb = new SmppByteBuffer(bytes);
         try {
-            short b = bb.removeByte();
+            int b = bb.removeByte();
             for (TON ton : TON.values()) {
                 if (ton.getValue() == b) {
                     destAddrTon = ton;
