+++ src/main/java/org/bulatnig/smpp/pdu/tlv/AlertOnMessageDelivery.java
@@ -33,7 +33,7 @@ public class AlertOnMessageDelivery extends TLV {
     }
 
     @Override
-    protected void parseValue(byte[] bytes, final EsmClass esmClass, final short dataCoding) throws TLVException {
+    protected void parseValue(byte[] bytes, final EsmClass esmClass, final int dataCoding) throws TLVException {
         if (getTag() != ParameterTag.ALERT_ON_MESSAGE_DELIVERY) {
             throw new ClassCastException();
         }
