+++ src/main/java/org/bulatnig/smpp/pdu/tlv/DestinationPort.java
@@ -41,7 +41,7 @@ public class DestinationPort extends TLV {
     }
 
     @Override
-    protected void parseValue(byte[] bytes, final EsmClass esmClass, final short dataCoding) throws TLVException {
+    protected void parseValue(byte[] bytes, final EsmClass esmClass, final int dataCoding) throws TLVException {
         if (getTag() != ParameterTag.DESTINATION_PORT) {
             throw new ClassCastException();
         }
