+++ src/main/java/org/bulatnig/smpp/pdu/tlv/ReceiptedMessageId.java
@@ -48,7 +48,7 @@ public class ReceiptedMessageId extends TLV {
     }
 
     @Override
-    protected void parseValue(byte[] bytes, final EsmClass esmClass, final short dataCoding) throws TLVException {
+    protected void parseValue(byte[] bytes, final EsmClass esmClass, final int dataCoding) throws TLVException {
         if (getTag() != ParameterTag.RECEIPTED_MESSAGE_ID) {
             throw new ClassCastException();
         }
