+++ src/main/java/org/bulatnig/smpp/pdu/tlv/SourceSubaddress.java
@@ -50,7 +50,7 @@ public class SourceSubaddress extends TLV {
     }
 
     @Override
-    protected void parseValue(byte[] bytes, final EsmClass esmClass, final short dataCoding) throws TLVException {
+    protected void parseValue(byte[] bytes, final EsmClass esmClass, final int dataCoding) throws TLVException {
         if (getTag() != ParameterTag.SOURCE_SUBADDRESS) {
             throw new ClassCastException();
         }
