+++ src/main/java/org/bulatnig/smpp/pdu/udh/UDHType.java
@@ -8,28 +8,28 @@ package org.bulatnig.smpp.pdu.udh;
  * Time: 19:37:12
  */
 public enum UDHType {
-    IEI((short) 0x00),
-    FORMATTED_TEXT((short) 0x0A),
-    PREDEFINED_MELODY((short) 0x0B),
-    LOADABLE_MELODY((short) 0x0C),
-    ANIMATED_CARTOON_LARGE((short) 0x0E),
-    ANIMATED_CARTOON((short) 0x0F),
-    LOADABLE_PICTURE((short) 0x11),
-    UPI((short) 0x13),
-    COLOR_PICTURE((short) 0x14),
-    ODI((short) 0x17);
+    IEI(0x00),
+    FORMATTED_TEXT(0x0A),
+    PREDEFINED_MELODY(0x0B),
+    LOADABLE_MELODY(0x0C),
+    ANIMATED_CARTOON_LARGE(0x0E),
+    ANIMATED_CARTOON(0x0F),
+    LOADABLE_PICTURE(0x11),
+    UPI(0x13),
+    COLOR_PICTURE(0x14),
+    ODI(0x17);
 
     /**
      * Численное выражение ID команды.
      */
-    private short value;
+    private int value;
 
     /**
      * Конструктор.
      *
      * @param shortVal численное выражение ID команды
      */
-    private UDHType(final short shortVal) {
+    private UDHType(final int shortVal) {
         value = shortVal;
     }
 
@@ -38,7 +38,7 @@ public enum UDHType {
      *
      * @return число
      */
-    public final short getValue() {
+    public final int getValue() {
         return value;
 	}
